****[REVERT_ALL]****
수정하는 것 만으로는 리뷰를 모두 반영하기 어렵다고 판단
처음부터 바우처 프로그램 재구성할 예정
