+++ .gitignore
@@ -1,11 +1,3 @@
-# Created by https://www.toptal.com/developers/gitignore/api/java,intellij,gradle
-# Edit at https://www.toptal.com/developers/gitignore?templates=java,intellij,gradle
-
-### Intellij ###
-# Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
-# Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
-
-.idea
 
 # CMake
 cmake-build-*/
