--- src/main/java/org/programmers/spbw1/io/Input.java
@@ -1,7 +1,0 @@
-package org.programmers.spbw1.io;
-
-import java.io.IOException;
-
-public interface Input {
-    String input(String prompt) throws IOException;
-}
