+++ src/main/java/org/programmers/voucherProgram/VoucherProgram.java
@@ -1,0 +1,27 @@
+package org.programmers.voucherProgram;
+
+import org.programmers.voucherProgram.io.Input;
+import org.programmers.voucherProgram.io.Output;
+
+import org.programmers.voucherProgram.voucher.Model.VoucherType;
+import org.programmers.voucherProgram.voucher.Model.Voucher;
+import org.programmers.voucherProgram.voucher.Repository.VoucherRepository;
+import org.programmers.voucherProgram.voucher.service.VoucherService;
+import org.programmers.voucherProgram.voucher.validator.VoucherValidator;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.lang.reflect.Method;
+import java.util.Optional;
+
+public class VoucherProgram implements Runnable {
+
+    @Override
+    public void run() {
+    }
+    private void create(){
+    }
+
+    private void list(){
+    }
+}
