+++ src/main/java/org/programmers/voucherProgram/io/Input.java
@@ -1,0 +1,6 @@
+package org.programmers.voucherProgram.io;
+
+import java.io.IOException;
+
+public interface Input {
+}
