+++ src/main/java/org/programmers/voucherProgram/io/Output.java
@@ -1,0 +1,6 @@
+package org.programmers.voucherProgram.io;
+
+import org.programmers.voucherProgram.voucher.Repository.VoucherRepository;
+
+public interface Output {
+}
