+++ src/main/java/org/programmers/voucherProgram/voucher/Model/PercentDiscountVoucher.java
@@ -1,0 +1,7 @@
+package org.programmers.voucherProgram.voucher.Model;
+
+import java.util.UUID;
+
+
+public class PercentDiscountVoucher extends Voucher{
+}
