+++ src/main/java/org/programmers/voucherProgram/voucher/Model/Voucher.java
@@ -1,0 +1,6 @@
+package org.programmers.voucherProgram.voucher.Model;
+
+import java.util.UUID;
+
+public abstract class Voucher {
+}
