+++ src/main/java/org/programmers/voucherProgram/voucher/Repository/MemoryVoucherRepository.java
@@ -1,0 +1,11 @@
+package org.programmers.voucherProgram.voucher.Repository;
+
+import org.programmers.voucherProgram.voucher.Model.Voucher;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+// @Repository
+public class MemoryVoucherRepository extends VoucherRepository{
+}
