+++ src/main/java/org/programmers/voucherProgram/voucher/Repository/VoucherRepository.java
@@ -1,0 +1,11 @@
+package org.programmers.voucherProgram.voucher.Repository;
+
+import org.programmers.voucherProgram.voucher.Model.Voucher;
+
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+
+public abstract class VoucherRepository {
+}
