+++ src/main/java/org/programmers/voucherProgram/voucher/service/VoucherService.java
@@ -1,0 +1,15 @@
+package org.programmers.voucherProgram.voucher.service;
+
+import org.programmers.voucherProgram.voucher.Model.FixedAmountVoucher;
+import org.programmers.voucherProgram.voucher.Model.VoucherType;
+import org.programmers.voucherProgram.voucher.Model.PercentDiscountVoucher;
+import org.programmers.voucherProgram.voucher.Repository.VoucherRepository;
+import org.programmers.voucherProgram.voucher.Model.Voucher;
+import org.springframework.stereotype.Service;
+
+import java.text.MessageFormat;
+import java.util.UUID;
+
+// @Service
+public class VoucherService {
+}
