+++ src/main/java/org/programmers/voucherProgram/voucher/validator/VoucherValidator.java
@@ -1,0 +1,11 @@
+package org.programmers.voucherProgram.voucher.validator;
+
+import org.programmers.voucherProgram.voucher.Model.VoucherType;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class VoucherValidator {
+    public Long validate() throws NumberFormatException{
+        return 0L;
+    }
+}
